+++ Source/VolumeSlider.cpp
@@ -12,10 +12,10 @@ VolumeSlider::VolumeSlider()
 void VolumeSlider::paint(Graphics& g)
 {
 	const int indent = LookAndFeel::getDefaultLookAndFeel().getSliderThumbRadius(*this);
-	double zeroPosition = (1.0 - valueToProportionOfLength(1.0)) * (getHeight() - 2 * indent);
+	float zeroPosition = static_cast<float>((1.0 - valueToProportionOfLength(1.0)) * (getHeight() - 2 * indent));
 
 	g.setColour(Colours::black);
-	g.drawLine(0, indent + zeroPosition, getWidth(), indent + zeroPosition);
+	g.drawLine(0.0f, indent + zeroPosition, static_cast<float>(getWidth()), indent + zeroPosition);
 
 	Slider::paint(g);
 }
